Add an implicits package.
Makes it easier to get all the instances in scope, following Cats’ approach.
