+++ core/shared/src/main/scala/matryoshka/Embed.scala
@@ -16,7 +16,7 @@
 
 package matryoshka
 
-import matryoshka.Recursive.ops._
+import matryoshka.implicits._
 
 import scala.Some
