+++ core/shared/src/main/scala/matryoshka/patterns/package.scala
@@ -16,6 +16,8 @@
 
 package matryoshka
 
+import matryoshka.implicits._
+
 import scala.Option
 
 import scalaz._, Scalaz._, Leibniz._
