+++ tests/src/test/scala/matryoshka/instances/fixedpoint/nat.scala
@@ -16,9 +16,10 @@
 
 package matryoshka.instances.fixedpoint
 
-import matryoshka._, Recursive.ops._
+import matryoshka._
 import matryoshka.data.Mu
 import matryoshka.helpers._
+import matryoshka.implicits._
 import matryoshka.scalacheck.arbitrary._
 
 import scala.Option
