+++ tests/src/test/scala/matryoshka/instances/fixedpoint/partial.scala
@@ -16,7 +16,8 @@
 
 package matryoshka.instances.fixedpoint
 
-import matryoshka._, Recursive.ops._, FunctorT.ops._
+import matryoshka._
+import matryoshka.implicits._
 import matryoshka.scalacheck.arbitrary._
 import matryoshka.specs2.scalacheck.CheckAll
