+++ tests/src/test/scala/matryoshka/patterns/diff.scala
@@ -16,11 +16,12 @@
 
 package matryoshka.patterns
 
-import matryoshka._, Recursive.ops._
+import matryoshka._
 import matryoshka.data.Mu
 import matryoshka.example._
 import matryoshka.exp._
 import matryoshka.helpers._
+import matryoshka.implicits._
 import matryoshka.scalacheck.arbitrary._
 import matryoshka.specs2.scalacheck.CheckAll
